requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Video Download
yt-dlp>=2024.1.0

# Social Media Uploaders
instagrapi>=2.0.0

# AI
google-genai>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Config
pydantic-settings>=2.0.0

